// frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseURL: `${string}://${string}` | (string & {});
};

/**
 * AIRequest
 */
export type AiRequest = {
    /**
     * Prompt
     *
     * The user prompt
     */
    prompt: string;
};

/**
 * AIResponse
 */
export type AiResponse = {
    /**
     * Answer
     *
     * The generated answer
     */
    answer: string;
    /**
     * Confidence
     *
     * Confidence score 0-1
     */
    confidence: number;
    /**
     * Sources
     *
     * Referenced sources if any
     */
    sources?: Array<string>;
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
    /**
     * Input
     */
    input?: unknown;
    /**
     * Context
     */
    ctx?: {
        [key: string]: unknown;
    };
};

export type UtilsHealthCheckData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/utils/health-check/';
};

export type UtilsHealthCheckResponses = {
    /**
     * Response Utils-Health Check
     *
     * Successful Response
     */
    200: boolean;
};

export type UtilsHealthCheckResponse = UtilsHealthCheckResponses[keyof UtilsHealthCheckResponses];

export type AiGenerateTextData = {
    body: AiRequest;
    path?: never;
    query?: never;
    url: '/api/v1/ai/generate';
};

export type AiGenerateTextErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type AiGenerateTextError = AiGenerateTextErrors[keyof AiGenerateTextErrors];

export type AiGenerateTextResponses = {
    /**
     * Successful Response
     */
    200: AiResponse;
};

export type AiGenerateTextResponse = AiGenerateTextResponses[keyof AiGenerateTextResponses];